Restructure validate package
